news-archiver/requirements.txt
requests
ollama
python-dotenv
notion-client
PyGithub